==> sc RNA-seq/step13-cellchat.R <==
###cancer cell和TME的细胞间相互作用
#思路：从上皮亚群中提取出来cancer，TME可以选取一些有代表性的
rm(list=ls())
library(Seurat)
options(stringsAsFactors = F)
library(SeuratObject)
library(ggplot2)
library(clustree)
library(cowplot)
library(dplyr)
library(CellChat)
getwd()
dir.create("13-cellchat")
setwd('13-cellchat/')
Tsce=readRDS( "../12-T/Tsce_celltype.rds")
myeloidsce=readRDS( "../12-T/myeloidsce_celltype.rds")
epi = readRDS( "../7-epi/epi_sce_celltype.rds")
table(epi$celltype)
cancer = epi[, epi$celltype %in% c( 'Cancer(CRABP2+)','Cancer(TM4SF1+)','Cancer(UBE2C+)', 'Clara-like cancer')]

sce = merge(Tsce,list(myeloidsce,cancer),
            add.cell.ids = c('T','myeloid','cancer'))

table(sce$sample)
table(sce$celltype)
Idents(sce) = sce$celltype
colnames(sce@meta.data)

names(sce@assays$RNA@layers)
sce[["RNA"]]$counts 
# Alternate accessor function with the same result
LayerData(sce, assay = "RNA", layer = "counts")
sce
sce <- JoinLayers(sce)
sce
dim(sce[["RNA"]]$counts )
AIS = sce[, sce$sample %in% c( 'AIS')]
MIA = sce[, sce$sample %in% c( 'MIA')]
IAC = sce[, sce$sample %in% c( 'IAC')]
# 
# sce=CreateSeuratObject(counts = sce@assays$RNA$counts,
#                        meta.data = sce@meta.data )
# sce

save(sce,file = 'input.Rdata')

cellchatA = function(sce){
  #创建cellchat对象
  cellchat <- createCellChat(sce@assays$RNA$data, meta = sce@meta.data, group.by = "celltype")
  levels(cellchat@idents)
  groupSize <- as.numeric(table(cellchat@idents)) 
  CellChatDB <- CellChatDB.human#导入配受体库
  showDatabaseCategory(CellChatDB) #查看描述该数据库组成的饼状图
  #直接使用CellChatDB全库进行细胞通讯分析：
  ##CellChatDB.use <- CellChatDB 
  CellChatDB.use <- subsetDB(CellChatDB, search = "Secreted Signaling")#选择特定的信号来进行分析，这里还可以选择ECM-receptor和Cell-Cell Contact。 
  cellchat@DB <- CellChatDB.use
  
  #预处理表达数据以进行细胞间通讯分析
  # subset the expression data of signaling genes for saving computation cost
  cellchat <- subsetData(cellchat)
  cellchat <- identifyOverExpressedGenes(cellchat)
  cellchat <- identifyOverExpressedInteractions(cellchat)
  # project gene expression data onto PPI network (optional)
  cellchat <- projectData(cellchat, PPI.human) #PPI.human PPI.mouse
  
  
  ###2.细胞通讯预测
  #计算通信概率并推断通信网络
  #Compute the communication probability and iC3D1er cellular communication network
  cellchat <- computeCommunProb(cellchat)
  # Filter out the cell-cell communication if there are only few number of cells in certain cell groups
  #这里还是不要过滤了，不然有些细胞亚群细胞数量过少被过滤掉之后会造成数据不一致而无法进行后续分析
  #cellchat <- filterCommunication(cellchat, min.cells = 10)
  #2）提取配受体对细胞通讯结果表：
  df.net <- subsetCommunication(cellchat, slot.name = 'net')
  head(df.net) #得到配受体对细胞通讯结果表
  # #或访问其它感兴趣/特定的细胞通讯结果：
  # df.net1 <- subsetCommunication(cellchat,
  #                                sources.use = c('LC'),
  #                                targets.use = c('FBN1+ FIB')) #访问特定细胞对子集
  # head(df.net1)
  df.net2 <- subsetCommunication(cellchat, signaling = c('MIF')) #访问特定信号通路子集
  head(df.net2)
  #3）提取信号通路水平的细胞通讯表：
  cellchat <- computeCommunProbPathway(cellchat) #计算信号通路水平上的通讯概率
  df.netp <- subsetCommunication(cellchat, slot.name = 'netP') #得到信号通路水平细胞通讯表
  head(df.netp)
  
  #4）细胞互作关系展示：
  #计算细胞对间通讯的数量和概率强度
  cellchat <- aggregateNet(cellchat)
  return(cellchat)
} 

IACcellchat = cellchatA(IAC)
MIAcellchat = cellchatA(MIA)
AIScellchat = cellchatA(AIS)


object.list <- list(AIS = AIScellchat,MIA = MIAcellchat,IAC = IACcellchat )
cellchat <- mergeCellChat(object.list, add.names = names(object.list), cell.prefix = TRUE)

gg1 <- compareInteractions(cellchat, show.legend = F, group = c(1,2,3), measure = "count")
gg2 <- compareInteractions(cellchat, show.legend = F, group = c(1,2,3), measure = "weight")
p <- gg1 + gg2
p
ggsave("Overview_number_strength.pdf", p, width = 10, height = 7)

#红色是case相对于control上调的，蓝色是下调的。
par(mfrow = c(1,3))
netVisual_diffInteraction(cellchat, weight.scale = T)
netVisual_diffInteraction(cellchat, weight.scale = T, measure = "weight")


#细胞互作数量对比网络图
par(mfrow = c(1,3))
weight.max <- getMaxWeight(object.list, attribute = c("idents","count"))
for (i in 1:length(object.list)) {
  netVisual_circle(object.list[[i]]@net$count, weight.scale = T, label.edge= F, 
                   edge.weight.max = weight.max[2], edge.width.max = 12, 
                   title.name = paste0("Number of interactions - ", names(object.list)[i]))
}
#左图是control，右图是case，可以直接对比数量变化。


#3.3 保守和特异性信号通路的识别与可视化
## 通路信号强度对比分析
#comparison可以修改
gg1 <- rankNet(cellchat, mode = "comparison", stacked = T, do.stat = TRUE)
gg2 <- rankNet(cellchat, mode = "comparison", stacked = F, do.stat = TRUE)
p <- gg1 + gg2
p
ggsave("Compare_pathway_strengh.pdf", p, width = 10, height = 6)
#左图最下面5个信号通路是case组独有的


pathways.show <- c("TGFb") 
weight.max <- getMaxWeight(object.list, slot.name = c("netP"), attribute = pathways.show) 
par(mfrow = c(1,3), xpd=TRUE)
for (i in 1:length(object.list)) {
  netVisual_aggregate(object.list[[i]], signaling = pathways.show, layout = "circle", 
                      edge.weight.max = weight.max[1], edge.width.max = 10, 
                      signaling.name = paste(pathways.show, names(object.list)[i]))
}


#热图
par(mfrow = c(1,3), xpd=TRUE)
ht <- list()
for (i in 1:length(object.list)) {
  ht[[i]] <- netVisual_heatmap(object.list[[i]], signaling = pathways.show, color.heatmap = "Reds",
                               title.name = paste(pathways.show, "signaling ",names(object.list)[i]))
}
ComplexHeatmap::draw(ht[[1]] + ht[[2]]+ ht[[3]], ht_gap = unit(0.5, "cm"))
